chore: P2 CI gate, year-state cleanup, rename activitiesData (SHA-1)
1. Frontend CI workflow runs pnpm ci on PRs touching frontend paths.
2. Dependabot groups react / react-dom / @types/* to avoid version skew.
3. Remove year-change useEffect setState (clear in changeYear handler).
4. Rename hooks/useActivities → data/activitiesData (module, not a hook).
