+++ src/components/LocationStat/CitiesStat.tsx
@@ -1,5 +1,5 @@
 import Stat from '@/components/Stat'
-import activitiesData from '@/hooks/useActivities'
+import activitiesData from '@/data/activitiesData'
 
 // only support China for now
 function CitiesStat({ onClick }: { onClick: (_city: string) => void }) {
